WJPr: World Justice Project Data and Visualization Tools

The WJPr package contains the most important data and visualization tools used by Data Analytics Unit at The World Justice Project. Version 1.0.0 comes with a wide range of visualization functions for replicating the image charts from the Country Reports. Additionally, you can find the Rule of Law Index scores data for all factors and subfactors as complementary data.

See also

Author

Maintainer: Carlos A. Toruno Paniagua ctoruno@worldjusticeproject.org

Authors:

World Justice Project Data Analytics Unit data-analytics@worldjusticeproject.org [copyright holder]