GPP Sample Data

A subset of data from the original General Population Poll Data. Countries and years have been renamed and grouped into three fictional countries (Atlantis, Narnia, Neverland) and multiple years.

Usage

gpp

Format

gpp

A data frame with 750 rows and 19 columns:

country

Country name (Atlantis, Narnia, or Neverland)

year

Survey year (2017, 2019, 2022)

gend

Gender (1 = Male, 2 = Female)

age

Age of respondent

q1a

Trust in Institution A (1-4 scale, 99 = Don't know)

q1b

Trust in Institution B (1-4 scale, 99 = Don't know)

q1c

Trust in Institution C (1-4 scale, 99 = Don't know)

q1d

Trust in Institution D (1-4 scale, 99 = Don't know)

q49a

Justice System evaluation A (1-4 scale, 99 = Don't know)

q49b

Justice System evaluation B (1-4 scale, 99 = Don't know)

q49c

Justice System evaluation C (1-4 scale, 99 = Don't know)

q49d

Justice System evaluation D (1-4 scale, 99 = Don't know)

q49e

Justice System evaluation E (1-4 scale, 99 = Don't know)

q49f

Justice System evaluation F (1-4 scale, 99 = Don't know)

q49g

Justice System evaluation G (1-4 scale, 99 = Don't know)

q49h

Justice System evaluation H (1-4 scale, 99 = Don't know)

q49i

Justice System evaluation I (1-4 scale, 99 = Don't know)

Source

https://worldjusticeproject.org/