Plot a Dots Chart following WJP style guidelines

wjp_dotsChart() takes a data frame with a specific data structure (usually long shaped) and returns a ggplot object with a dots chart following WJP style guidelines.

Usage

wjp_dotsChart(
  data,
  target,
  grouping,
  labels,
  cvec = NULL,
  order,
  diffOpac = F,
  opacities = NULL,
  diffShp = F,
  shapes = NA,
  ptheme = WJP_theme()
)

Arguments

data

Data frame containing the data to plot

target

String. Column name of the variable that will supply the values to plot.

grouping

String. Column name of the variable that supplies the grouping values. The plot will show a different color per group.

labels

String. Column name of the variable that supplies the Y-Axis labels to show in the plot.

cvec

Named vector with the colors to apply to the dots. Default is NULL.

order

String. Column name of the variable that contains the custom order for the labels.

diffOpac

Boolean. If TRUE, the plot will expect different levels of opacities for the dots. Default is FALSE.

opacities

Named vector with the opacity levels to apply to the dots. Default is NULL.

diffShp

Boolean. If TRUE, the plot will expect different shapes for the dots. Default is FALSE.

shapes

Named vector with shapes to be displayed. Default is NULL.

ptheme

ggplot theme function to apply to the plot. By default, function applies WJP_theme().

Value

A ggplot object

Examples

if (FALSE) {
# Always load the WJP fonts if not passing a custom theme to function
wjp_fonts()

# Preparing data
data2plot <- gpp %>% 
select(country, q49a)%>%
  mutate(div = case_when(
    q49a == 1  ~ "Very confident",
    q49a == 2  ~ "Fairly confident",
    q49a == 3  ~ "Not very confident",
    q49a == 4  ~ "Not at all confident",
    q49a == 99 ~ NA_character_
  )
  )%>% drop_na()%>%
  group_by(country, div)%>%
  count()%>%
  group_by(country)%>%
  transmute(div, value2plot = n/sum(n)*100)%>%
  mutate(order_value = case_when(
    country == "Atlantis" ~ 1,
    country == "Narnia" ~ 2,
    country == "Neverland" ~ 3))

# Plotting chart
wjp_dotsChart(data= data2plot, target = "value2plot", grouping = "div", order = "order_value", labels = "country")
}