Plot a Dumbbell Chart following WJP style guidelines

wjp_dumbbells() takes a data frame with a specific data structure (usually long shaped) and returns a ggplot object with a dumbbell chart following WJP style guidelines.

Usage

wjp_dumbbells(
  data,
  target,
  rows,
  color,
  cgroups,
  labels = NULL,
  labpos = NULL,
  cvec = NULL,
  order = NULL,
  ptheme = WJP_theme()
)

Arguments

data

A data frame containing the data to be plotted.

target

A string specifying the variable in the data frame that contains the numeric values to be plotted.

rows

A string specifying the variable in the data frame that contains the categories for the rows.

color

A string specifying the variable in the data frame that indicates the groups for start and end points.

cgroups

A vector of two strings specifying the groups to be compared in the dumbbell plot.

labels

A string specifying the variable in the data frame that contains the text labels to display.

labpos

A string specifying the variable in the data frame that contains the label positions.

cvec

A vector of colors to apply to the points and lines. Default is NULL.

order

A named vector specifying the order of the categories. Default is NULL.

ptheme

A ggplot2 theme object to be applied to the plot. Default is WJP_theme().

Value

A ggplot object representing the dumbbell plot.

Examples

if (FALSE) { # \dontrun{
data <- data.frame(
  group = c("A", "B", "C", "A", "B", "C"),
  target = c(30, 50, 20, 40, 60, 30),
  color = c("start", "start", "start", "end", "end", "end")
)
cgroups <- c("start", "end")
cvec <- c("#2a2a94", "#d1cfd1")

wjp_dumbbells(data, "target", "group", "color", cgroups, cvec)
} # }