Plot a Horizontal Edgebars Chart following WJP style guidelines

wjp_edgebars() takes a data frame with a specific data structure (usually long shaped) and returns a ggplot object with an edgebar chart following WJP style guidelines.

Usage

wjp_edgebars(
  data,
  y_value,
  x_var,
  label_var,
  color_var = NULL,
  bar_colors = NULL,
  x_lab_pos = NULL,
  y_lab_pos = 0,
  nudge_lab = 2.5,
  margin_top = 20,
  bar_width = 0.35,
  ptheme = WJP_theme()
)

Arguments

data

A data frame containing the data to be plotted.

y_value

A string specifying the variable in the data frame that contains the numeric values to be plotted as bars.

x_var

A string specifying the variable in the data frame that contains the categories for the x-axis.

label_var

A string specifying the variable in the data frame that contains the labels to be displayed near the bars.

color_var

A string specifying the variable in the data frame that contains the the color groupings for the bars. Default is NULL.

bar_colors

A named vector specifying the colors for the bars. Default is NULL.

x_lab_pos

A string specifying the variable in the data frame that contains the order in which the bars will be displayed. Default is NULL.

y_lab_pos

A numeric value specifying the y-axis position for displaying labels. Default is 0.

nudge_lab

A numeric value specifying the padding for displaying labels in milimeters. Default is 2.5.

margin_top

A numeric value specifying the top margin of the plot. Default is 20.

bar_width

A numeric value specifying the width of the bars. For single bars the default value of 0.35 is recommended, for plots with two bars a value of 0.5 is more suitable.

ptheme

A ggplot aesthetic theme to be applied to the chart. Default is the WJP_theme initilized with the package.

Value

A ggplot object representing the edge bars plot.

Examples

if (FALSE) { # \dontrun{
data <- data.frame(
  y_value = c(30, 50, 20),
  x_var = c("Category A", "Category B", "Category C"),
  label_var = c("Label A", "Label B", "Label C")
)

wjp_edgebars(data, "y_value", "x_var", "label_var")
} # }