Plot a Radar Chart following WJP style guidelines

#' @description wjp_radar() takes a data frame with a specific data structure (usually long shaped) and returns a ggplot object with a radar chart following WJP style guidelines.

Usage

wjp_radar(
  data,
  axis_var,
  target_var,
  label_var,
  color_var,
  colors,
  order_var,
  maincat,
  source = "GPP"
)

Arguments

data

A data frame containing the data to be plotted.

axis_var

A string specifying the variable in the data frame that contains the groups for the axis.

target_var

A string specifying the variable in the data frame that contains the values to be plotted.

label_var

A string specifying the variable in the data frame that contains the labels to be displayed.

color_var

A string specifying the variable in the data frame that contains the color groupings.

colors

A vector of colors to apply to lines. The first color will be applied to percentages in labels.

order_var

A string specifying the variable in the data frame that contains the display order of categories.

maincat

A string indicating the category of labels to show in the radar.

source

A string which can take two values (GPP or QRQ).

Value

A ggplot object representing the radar plot.

Examples

if (FALSE) { # \dontrun{
data <- data.frame(
  axis_var = rep(letters[1:5], each = 3),
  target_var = runif(15, 0, 1),
  label_var = rep(LETTERS[1:5], each = 3),
  order_var = rep(1:5, each = 3),
  year = rep(2020:2022, 5)
)
colors <- c("red", "blue", "green")
maincat <- 2022

wjp_radar(data, "axis_var", "target_var", "label_var", "order_var", colors, maincat)
} # }